+++ tests/tests.py
@@ -15,13 +15,14 @@ class BotTest(unittest.IsolatedAsyncioTestCase):
         self.assertEqual(bot_req(), {"status": 400, "ret": "Invalid JSON"})
         self.assertEqual(
             bot_req({"toto": 3}),
-            {"status": 400, "ret": "Missing body and/or API key property"},
+            {"status": 400, "ret": "Missing body, key, room_id"},
         )
         self.assertEqual(
-            bot_req({"body": 3}, "wrong_key"), {"status": 401, "ret": "Invalid API key"}
+            bot_req({"body": 3}, "wrong_key", "wrong_room"),
+            {"status": 401, "ret": "Invalid API key"},
         )
         self.assertEqual(
-            bot_req({"body": 3}, "wrong_key", key_as_param=True),
+            bot_req({"body": 3}, "wrong_key", "wrong_room", key_as_param=True),
             {"status": 401, "ret": "Invalid API key"},
         )
         self.assertEqual(
@@ -31,10 +32,11 @@ class BotTest(unittest.IsolatedAsyncioTestCase):
         # TODO: if the client from matrix_webhook has olm support, this won't be a 403 from synapse,
         # but a LocalProtocolError from matrix_webhook
         self.assertEqual(
-            bot_req({"body": 3}, KEY), {"status": 403, "ret": "Unknown room"}
+            bot_req({"body": 3}, KEY, "wrong_room"),
+            {"status": 403, "ret": "Unknown room"},
         )
         self.assertEqual(
-            bot_req({"body": 3}, KEY, key_as_param=True),
+            bot_req({"body": 3}, KEY, "wrong_room", key_as_param=True),
             {"status": 403, "ret": "Unknown room"},
         )
 
@@ -58,6 +60,52 @@ class BotTest(unittest.IsolatedAsyncioTestCase):
         message = messages.chunk[0]
         self.assertEqual(message.sender, FULL_ID)
         self.assertEqual(message.body, text)
+        self.assertEqual(message.formatted_body, "<h1>Hello</h1>")
+
+    async def test_room_id_req(self):
+        """Send a markdown message in a room given as data, and check the result."""
+        body = "# Hello"
+        messages = []
+        client = nio.AsyncClient(MATRIX_URL, MATRIX_ID)
+
+        await client.login(MATRIX_PW)
+        room = await client.room_create()
+
+        self.assertEqual(
+            bot_req({"body": body, "room_id": room.room_id}, KEY, room.room_id),
+            {"status": 200, "ret": "OK"},
+        )
+
+        sync = await client.sync()
+        messages = await client.room_messages(room.room_id, sync.next_batch)
+        await client.close()
+
+        message = messages.chunk[0]
+        self.assertEqual(message.sender, FULL_ID)
+        self.assertEqual(message.body, body)
+        self.assertEqual(message.formatted_body, "<h1>Hello</h1>")
+
+    async def test_room_id_parameter(self):
+        """Send a markdown message in a room given as parameter, and check the result."""
+        body = "# Hello"
+        messages = []
+        client = nio.AsyncClient(MATRIX_URL, MATRIX_ID)
+
+        await client.login(MATRIX_PW)
+        room = await client.room_create()
+
+        self.assertEqual(
+            bot_req({"body": body}, KEY, room.room_id, room_as_parameter=True),
+            {"status": 200, "ret": "OK"},
+        )
+
+        sync = await client.sync()
+        messages = await client.room_messages(room.room_id, sync.next_batch)
+        await client.close()
+
+        message = messages.chunk[0]
+        self.assertEqual(message.sender, FULL_ID)
+        self.assertEqual(message.body, body)
         self.assertEqual(message.formatted_body, "<h1>Hello</h1>")
 
     async def test_markdown_body(self):